refactor(capabilities): extract control plane as wasm component supervisor (issue 603)
Phase 1 of issue 603. Substrate's `ControlPlane` + `Scheduler` collapse
into one cap, `aether-capabilities::ControlPlaneCapability`, that owns
the wasm component table, per-component dispatcher threads, and the
wasmtime Engine/Linker outright. The `Arc<RwLock<ComponentTable>>`
shared between substrate and ControlPlane retires; the cap installs
itself as the `Mailer`'s `ComponentRouter` at init and component-bound
mail routes through that interface.

The `#[actor]` macro now accepts mixed `#[handler]` + `#[fallback]`
(hybrid shape) — typed handlers stay authoritative at the type system,
the fallback runs only on dispatch-table misses for kinds that arrive
by mailbox name. The cap uses this for the Phase 1 chassis-peripheral
migration window (`capture_frame`, `set_window_mode`, `set_window_title`,
`platform_info` ride the fallback closure on `ControlPlaneConfig` until
Phases 2-4 peel them into their own caps).

Scenario runner's `Step::LoadComponent` now awaits `LoadResult` before
proceeding — pre-603 the legacy synchronous sink dispatch made
load-before-advance emergent; with the cap dispatching async on its
own thread, sequential semantics need explicit awaiting. Test bench's
`pump_until_reply` budget bumped from 1s to 10s to absorb wasm compile
under parallel-test contention.
