+++ crates/aether-capabilities/Cargo.toml
@@ -41,6 +41,7 @@ native = [
     "dep:tracing",
     "dep:ureq",
     "dep:url",
+    "dep:wasmtime",
 ]
 
 # Wasm-safe marker layers. Activate from a wasm component that wants to
@@ -77,6 +78,11 @@ postcard = { version = "1.1", optional = true, features = ["alloc"] }
 tracing = { version = "0.1", optional = true }
 ureq = { version = "3", optional = true, features = ["rustls"] }
 url = { version = "2", optional = true }
+# wasmtime is reached for by ControlPlaneCapability (issue 603) — the
+# cap holds the Engine/Linker/Module the substrate previously held in
+# its `ControlPlane` struct. Same `wasmtime = "30"` aether-substrate
+# uses; chassis builds get a single wasmtime in the dep tree.
+wasmtime = { version = "30", optional = true }
 wgpu = { version = "29.0.1", optional = true }
 png = { version = "0.17", optional = true }
 cpal = { version = "0.15", optional = true }
