+++ crates/aether-substrate/src/lib.rs
@@ -1,10 +1,19 @@
 //! aether-substrate: runtime that every substrate chassis shares.
 //!
-//! Hosts the wasmtime engine, the mail scheduler, the per-mailbox
-//! component table, the kind manifest, the reply-handle table, the
-//! control-plane dispatcher, and the hub-socket client. Chassis-specific
+//! Hosts the wasmtime engine, the mail router, the kind manifest, the
+//! reply-handle table, and the hub-socket client. Chassis-specific
 //! peripherals (window, GPU, TCP listener, event loop) live in the
 //! chassis crate that binds this as a dependency. See ADR-0035.
+//!
+//! The wasm-component supervisor — the actor that owns the per-mailbox
+//! component table, dispatcher threads, and the wasmtime Engine /
+//! Linker — lives in `aether-capabilities` as
+//! [`ControlPlaneCapability`][cp] (issue 603). Substrate exposes the
+//! interface the supervisor implements via [`supervisor::ComponentRouter`]
+//! plus the structured drain outcomes ([`supervisor::DrainSummary`])
+//! the chassis frame loop matches on; it knows nothing about the
+//! supervisor's identity beyond "something installed itself on
+//! [`Mailer::install_component_router`]".
 //!
 //! The `Chassis` trait (ADR-0035, redefined by ADR-0071) is universal
 //! but intentionally narrow: `const PROFILE` (the chassis's stable
@@ -13,17 +22,9 @@
 //! thread), `type Env` (resolved-config bag), and
 //! `fn build(env: Self::Env) -> Result<BuiltChassis<Self>, BootError>`.
 //! The chassis instance you `run()` is the [`BuiltChassis<Self>`] the
-//! trait method returns, not a value of `Self` itself. Chassis-specific
-//! control kinds (desktop's `capture_frame` / `set_window_mode` /
-//! `platform_info`, hub's future routing/operator kinds) ride through
-//! `ControlPlane::chassis_handler` — the fallback closure core's
-//! dispatch falls into for unknown kinds. That keeps any single
-//! chassis from having to implement `Unsupported` stubs for
-//! operations it doesn't support.
+//! trait method returns, not a value of `Self` itself.
 //!
-//! Helpers for chassis-side handlers live under `control`:
-//! `decode_payload` and `resolve_bundle` are pub so chassis dispatch
-//! can validate mail bundles the same way core does.
+//! [cp]: https://docs.rs/aether-capabilities/latest/aether_capabilities/struct.ControlPlaneCapability.html
 
 // Issue 552 stage 2: the `#[actor] impl NativeActor for X` macro
 // emits `impl ::aether_substrate::NativeDispatch for X` so external
@@ -41,7 +42,7 @@ pub mod capture;
 pub mod chassis;
 pub mod chassis_builder;
 pub mod component;
-pub mod control;
+pub mod control_helpers;
 pub mod ctx;
 pub mod frame_loop;
 pub mod handle_store;
@@ -60,10 +61,10 @@ pub mod registry;
 #[cfg(feature = "render")]
 pub mod render;
 pub mod reply_table;
-pub mod scheduler;
+pub mod supervisor;
 
 pub use aether_actor::Actor;
-pub use boot::{ChassisHandlerContext, SubstrateBoot, SubstrateBootBuilder};
+pub use boot::{SubstrateBoot, SubstrateBootBuilder};
 pub use capability::{
     ActorErased, BootError, BootedChassis, ChassisBuilder, ChassisCtx, DropOnShutdownClaim,
     Envelope, FallbackRouter, FrameBoundClaim, MailboxClaim, SinkSender, WedgedFrameBound,
@@ -74,7 +75,6 @@ pub use chassis_builder::{
     NeverDriver, NeverDriverRunning, NoDriver, PassiveChassis, RunError,
 };
 pub use component::Component;
-pub use control::{AETHER_CONTROL, ChassisControlHandler, ControlPlane};
 pub use ctx::SubstrateCtx;
 pub use input::{InputSubscribers, new_subscribers, remove_from_all, subscribers_for};
 pub use mail::{KindId, Mail, MailKind, MailboxId, ReplyTarget, ReplyTo};
@@ -86,7 +86,7 @@ pub use outbound::{
 };
 pub use panic_hook::init_panic_hook;
 pub use registry::{MailboxEntry, Registry, SinkHandler};
-pub use scheduler::Scheduler;
+pub use supervisor::{ComponentRouter, ComponentSendOutcome, DrainDeath, DrainOutcome, DrainSummary};
 
 /// Well-known mailbox name for substrate-level diagnostic events
 /// delivered back to this engine. Today the only kind delivered here
